Upload the address as StructuredPostalAddress
The address was not getting uploaded if we tried to dump the entire
address as a PostalAddress element. The only fix for the moment
appears to be to use a StructuredPostalAddress element. We will need
to breakout the PostalAddress into its components. This remains a
TODO. At least for now we are able to grab the full address and dump
it to Google.