Sync Logic - Part 1: Create Message lists on local side
We create a list of contacts in outlook that potentially need to be
synched to google contacts based on the last modified timestamp and
availability of the google contact id.

The modified list ar potential candidates for a "merge", whereas the
new lsi will be a straightforward copy to google.

There will be a similar set of lists in the other durection.