Remove pywin32 from tree and make it external dependency
There are some well known problems of using pywin32 in a portable
distribution. It is not worth the effort at this time to fight this battle.
It works quite fine if the correct pywin32 package is installed along with
python in the system wide site-packages directory.

Further, when we evenutually build a bundled distribution built with py2exe,
it is likely this will not be a problem for the junta. Developers can be
expected to install a few packages anyway...