Comment out 'Advanced Panel' and other unsupported features U
The current feature set only supports sync of contacts from
default store of default outlook profile. The UI had a lot of other
stuff for more advanced features that are not yet supported. In the
interest of shipping out a simple working beta lets comment out all the
UI cruft for now.