make ui.py the starting point.
tools/start_main.py was meant to be the main loop for the tornado server.
This was from a time when the UI was supposed to be a javascript
based page served by tornado. But we have since moved to a simple
wxPython interface contributed by consulite. With this change the
UI loop will anchor the app.

Should eventually nuke the tornado stuff from the repo.