Replace the OLID in Google instead of appending to existing OLIDs
After a successful copy to outlook we put a b64 encoded entryid to track
the sync status - as a user_defined_field. In case of synching modifications
we need to overwrite the old entryID - as our approach is to
delete the old outlook entry and to replace it with a brand new record.

This 'fix' is itself a hack and has a bug. In case there are other
user_define_field's this will fail to replace the old olid with the
correct new one. We need to fix this by first searching for the
correct olid and replacing it with the new one.