Mapi session logoff in a destructor
We are facing Outlook instability issues; After our program runs Outlook
fails to start unless the windows user logs off and logs back on.

One of the things that I tried to do which is a matter of good practice
anyway is to log off the mapi session before trying again. This does not
seem to work, actually. The main issue persists. But this code makes
sense anway. So let's just commit it.