First set of refactoring change sto support tasks sync
A bunch of commits here:

1. Separate out the Contact wrapper into a separate file ol_contact.py
2. Add a number of new classes to ol_wrapper.py to make it easier
   for working on multiple message stores inside a single outlook profile
3. Initial support for tasks and notes. Make a generic folder clas
   which can be iterated on.

These chnages will not work as is with the rest of the code and sync.
however these are important chages, and we are able to run teh
ol_wrapper.py to open the Outlook default profile and read the
three message stores that I have installed on my test machine.