Make all contacts operations on default contats folder
Recall that in the beginning of time Gout had all the operations
being performned on the default contacts folder, and this behaviour
was hard coded. Then we have been working on putting in a generic
infra to explore and work on any type of folder - tasks included.

Now that the required infrastructure is in place, this fix puts back
the default behaviour of synching the default contacts folder.

So essentially we are where we began at the start this branch -
in absolute feature terms - but with the code flexibility to look
at any of the folders.