File format version 2 of the app state config
The biggest change we are working on now is to support more than 2
types of PIM databases. This requires some fundamental changes in
how we capture the app configuraiton preferences as well as how
the db sync status is tracked by the application.

The app_state.json.example file, as always, is well documented
and contains enough comments to make the file self-explanatory

A number of general improvements have also been implemented in state.py
whereby range checks are done for some fields, etc.