Google Folder and Contact wrappers commit
With this commit the main derived classes GCContact and GCContactFolder are
defined, and useable to create and update contacts on the server. Even the
prep_gc_sync_lists() routine is working as it was before the rewrite.

This means that the Google Contacts side of things is about 75% ready for sync
stuff to be implemented.