Some infra for object attributes - over and above existing item properties
Differentite between item properties that get written to a datastore
and attributes that are only required for general book keeping for
the run time of the program. Other than that distinction these two
sets can be managed similarly. But for absolute clarity there are
now two separate and distinct APIs