Support for reading and writing back phone vectors from BBDB
There is an important change in how the phone numbers are managed,
and this impacts the other DBs as well. As BBDB supports
arbitrary labels for phone numbers, Each entry that is put
into the local represntation of vector using add_phone_home() is
now a tuple (label, number). This will have to be supported by
the other users of those routines.