Changes to _gc and _ol to deal with change in format of phone entries.
As per a recent change, all phone entires now retain their native labels.
With this change the contact_ol an contact_gc files implement this.

Note that Outlook does not a native representation of phone or email labels.
So a sync to Outlok and back will lose that label. This needs a bit more work
on the Outlook side and will need to be done separately.