implement del_folder in outlook - sort of
pywin32 does not support MAPI DeleteFolder() interface. As a result
we are unable to delete folders created automatically. This is not
a severe restriction as they can always be deleted from inside
MS Outlook.

But this commit contains code that will do the needful as and when
pywin32 supports the required interface to mapi.

For now the user is told that deletefolder is not possible, and
get on with life.