--op=del-profile implemented
implemented for all three pimdbs. Biggest change is the OL
backend, where the named properties used to store the
custom sync ids were statically decided and stored in the
proptags object. It is now dynamically created, based on the
profile-specified gid. The mapitag ASYNK_PR_GCID and
ASYNK_PR_BBID are now not used. Instead the values used are
exactly the same sync_labels as for the other pimdbs:
asynk:sample:gc type labels are used to lookup the property tag
values which are inturn used to look up sync tag properties from
MAPI.