Implement --op=clear-sync-artifacts operation
Now you can specify a folder-id, dbid, and a sync label regexp
to be cleared out of the system.

Note that if all is well and in steady state, the del-profile will
take care of cleaning up the sync artificats. However, if there
is a problem, or a bug, and sync got stuck in some form of
inconsistent state, the only way out could be clearing some
of the sync state manually and redo the operation. This particular
clear operation is meant to aid in this process.