Ensure that all sync methods return appropriate True or False values
The success and failure of sync operations will determine if the
timestamps are updated. We are being conservative here to ensure that
we do not lose any updates on unrelaed items if one or more items
cannot be synched. The error could happen in any path, including
writing back of sync tags, creation, updation, etc.