Set created() field for Google entries if not available
The created value is stored as a custom value. We do not get this
as a property from Google. A lack of a field here affects entries
synched to BBDB and Outlook. We settle for a compromise where we
just set the current time as the creation time - it will have the
effect that any contact copied to BBDB, say, will have a logial
and correct time as creation time. Since it is writte back to
Google as a custom property, nothing gets affected