"Folder" support for BBDB
BBDB does not support a built in folder system. It is all one flat
collection of contacts in the .bbdb. Here we try to logically
arrange the contacts collection into folders based on a notes field.

By default this notes field should have key named 'folder'. The value
of this field is taken as the name of the folder. The key name can
be changed by editing config.json in the appropriate place.

Note that the point of logically grouping the contacts into Folders
is to enable sync of only a subset of your BBDB contacts with an
external source. This is particularly helpful if you have a very large
collection of contacts, but are interested in only synching those with
phone numbers or from yoru company, or any other criterion you can think
of. Just add the folder note to those contacts, and use that folder name
in a new sync profile.