Fix the input parameter parsing and requirements to factor BBBD store/folder
We have changed BBDB structure sucht hat a 'store' is a bbbd file, and
a folder is an optional logical structure isnide the file specified
by the folder note field. This new structure needs follow on support
from parser, and the operation implementations.

With this fix, we now have --list-folders, --nw-folder --new-store and
--del-folder working as expected, even on BB. Code has been verified
with GC, where again, a concept of store has been impelemented, and
refers to the email address or the account login.