In outlook handle non-OL profile sync tags properly
This is turning out to be quite a tricky thing to get right. We
should leave alone any sync tag that is not related to the
current profile. It is turning out to be harder than we
expected.

There are problems in hte other PIMDB implementations that we
still need to address.