Custom Property support for Outlook
There is no straightforward way to stash away a number of fields
of custom fields in Outlook. Adding a single custom field is
very painful. To workaround, we will create a single custom field
and encode all the custom fields there as a json, which will be parsed
and populated into appropriate fields when there is a need.