Google Contacts Projection madness
full, thin, and base - these are the three things we have seen
in the 'projection' part of the contact URL. Each seems to have
problems. Worse, the Id in contact entries contain different things
at different times, including hte http field.

Here we try to perform a deterministic set of transofmraitons on
the GCID, the intent is that at every point where this GCID is
critical, we ensure the same format of ID is used.