Bugfix: Not all contact properties were being read from outlook
notes, given name, surnamne etc were getting missed. It is still not
clar what the problem was. Most likely the problem was we were
trying to get the contact properties based on a folder's default
columns. However what was bizzare is the same routine when
run from the test_ol routine was giving results that was not matching
from the regular sync flow.

At any rate, for now there is a solution, although I am less than
super confident.