Support for BBDB file format ver 6
With this fix, one should be able to use ASynK with BBDB v2
provided it is patched a bit. Details of this will be
inserted into the documentation.

This is not the cleanest solution. But we have anidea of
what to do in case the format changes again in BBDBv3.

Basically all the snarfing and tostring methods in contact_bb.py
need to reside in the message store class which is aware of
versions, etc.