Bugfix: Mobile and Work Emails multiply on repeated sync
Silly handling of a reference to a 'live' object, in this instance
the home phone number was getting munged with other data, and if
further sync operations happend with that object, it lead to
multiplying mobile and work numbers. The way the problem really
appeared totally put me off track... Very intersted.