Keep track in state.json of items in a collection at end of a sync run
This is important to identify the items that go missing, IOW deleted,
by the user between two sync runs. Google Contacts gives us a
way to identify all contacts that were deleted since a time stamp,
but we have no such luck with BBDB and Outlook, etc.

In subsequent work we will use this data to identify deletes and
take appropriate action.