Massage Outlook so as to sync Dates without year
Google Contacts allows the use of anniversary and birthdates
without an actual year value. That is not possible with Outlook
however. We work aroudn this by inserting a dummy year of 1887
and use it as a special value - adding it where Google does not
provide a year, and filter it out in the reverse direction.

To be able to do this we have to do some jumping through the
hoops:

a) Outlook returns values in UTC, so when synched dates appear to
   shift by one day

b) the 1887 year is way earlier than the epoch. So we have to deal
   with this year value with a bit of care.