Make ASynK installable in a central directory and have per-user config dirs
With this change, ASynK will try to use ~/.asynk/ as the configuration
directory to contain config.json and state.json. Also the logs and backups
directory will be stored here.

Note that this cannot be changed through a config.json variabel as this
directory itself is needed to locate config.json. The following
heuristic is used:

Asynk user directory is set to either ~/.asynk or the value of --user-dir
command line argument. If ~/.asynk/ is empty then the state.json and
config.json files are copied from the asynk application root directory.
This will help older users to transparently migrate their config and state
files to the new location without havign to do anything else.

Finally, made asynk.py executable so it can be executed directly.