Use sane defaults for store and folder ids if they are null in config
This is the final step needed to support a 'sane' sync option
with minimal configuration.

Basically, if a user has just fetched ASynK, the most likely common case
scenario will be to just sync his entire BBDB to the My Contacts folder
in Google Contacts, or defaults contacts folder in Outlook to Google.
With this commit that is possible.

The way this is enabeld is any gc sync profile without a folder id is
synhed to the My Contacts folder. If a store id is also missing a
google contacts username is prompted for, and stored in the default storeid
field. A similar "logical" thing is done with Outlook.