First element of email list in BBDB will be treated as the primary address
Gmail has a flag to identify the primary email address. In Emacs,
bbdb sets the first email address for tab completion and for completing
using the mail-alias field etc. So preserving this semantics is good.
This is what we acheive with this commit - ensure that the first
email is treated as the primary address - while reading as well as
when writing out to the BBDB database.