Stashing away code to track primary phone number
Similar to an earlier commit where we tracked the primary email
address and made it the first email in the BBDB list, we should be
tracking phone numbers as well. However the primary field for phone
numbers is not really set by Google properly (as best as I can see).

This diff is really an attempted patch which can come in handy if
and when we decide to explore this problem in more detail at some
point.