Proper handling of the Affix/Suffix field in BBDB
The Affix field was introduced in BBDBv3 and is an array of strings.
So far we have not really handled it too well, injecting illegal
fields into v6 bbdb files, and even in the best case ignoring
all but the first affix.

Now we remedy the situation, using the first affix value as the suffix
field, and stashing the rest away in custom field. When writing BBDB
we keep all affixes in the notes section in v6 while generating a
well formed affix field in v7.