Allow sync to non-existent BBDB Folders
With this fix, one can create a sync profile to an non-existant
bbdb folder, and it will spring to life when the actual sync
happens.

The actual use case could be something as follows:

You split your master Google Cotnacts list into "Personal" and
"Work", and would like to track it using simialr names in your
BBDB, then you would first create a new sync profile, and
ASynK will create contacts in the proper logical folder when the
contacts are written to disk.