pychecker catch: unused variables gid and statfi
there were two typos, one harmless and the other that could have
caused some painful bugs with Outlook syncs.

statefi should really be statei. Now fixed.

The other is more subtle. The routine that is used to compute the
olgid to be used at time of new profile creation works by first
making a list of all used gids for that dbid, and then finding
one that is unused. The array that was being built, due to the
typo, was an array of strings of hte profile names, instead of the
olgids of those profiles.

This bug will hurt if there are a number of profiles created and the
deleted, and then some more new ones are created. When this happens
the buggy code will check a olgid with a profile name, find that the
number is "unused" and then apply that for the new profile, leaing
to mixing of the contacts from different profiles.

Phew, good catch, pychecker.