pychecker catch: unused, weird code about sorted dbids in profile
My memory is really failing me here. pychecker has flagged two
instances where there are fragments of code and comments that suggest
the dbids need to be in sorted order before being operated on.
pychecker caught these snippets because they are unused code
chunks. I am not able to find any other surviving code that
depends on this sorted order. I am really unable to remember why
I needed this stuff, and when we move away from this.

due to this confusion, I am just commenting out the unsued code
in question along with some comments so we can come abck to this
stuff if required.