Bugfix: Ensure sync_dir option of SYNC1WAY does the right thing
Due to erroneos initialization of profile settings after reading
from config file / command line, a state.json configuration setting
was not working as expected when sycn_dir was set to SYNC1WAY. If hte
--sync-dir=1way was set, it should have worked as expected.

This is now fixed.

Noted that the same issue was present with the conflict_resolve
option. This is also fixed now.