Support ~/.netrc entries for specifying gmail passwords
Currently ASynK does not store any passwords (for your gmail profiles).
The user is prompted for the password at the right time, and the
password is echoed on the screen. It is done this way because
getpass does not work well with input redirection, so someone who
wants to avoid typing the password everytime faces an issue.

With this fix we leverage the age-old ~/.netrc file format to store
credentials needed by ASynK. Here is how it works:

- Add an entry in ~/.netrc with machine as 'gc_xxxx' where xxxx is the
  name of your sync profile

- For your gc_xxxx machine, add a login that is your gmail id

- Also give a password in cleartext that is your gmail password.

A sample entry will look like this:

machine gc_testbbgc
  login yyyy
  password xxxx

Here we have a sync profile called testbbgc, which is syching your
bbdb to yyyy@gmail.com and your password is xxxx

Note following behaviours:

- If your profile is configured with a differnet login id, then this
  ~/.netrc entry is ignored

- If you provide a password on the command line, this ~/.netrc entry
  is ignored

If you have no matching entry in netrc and nothing is provided on the
command line, then the current behaviour will be retained. (this can
now be changed to getpass...)