Workaround owncloud barfing on getProperties() on the root addressbook
For Apple Calendar Server the main addressbook was easily obtained
through a call to self.session().getProperties(URL(url=root), props)
However owncloud does not return anything for the above call. To
workaround this, we catch the condition when nothing is returned
and just lookup in the full folder list.