Clean up unwanted state.py helpers related to exgid
We put in a lot of code for exchange store profile management that
was copid from the outlook code. In particualr the use of the olgid
profile state flag was replicate. However we now do not need to use
that methodology at all. Recall that the olgid was used to create
a custom named integer extended property for each profile involving
outlook. In Exchange store we just put all of the sync tags as
part of a separate custom property that is a named string identified
extended property.

We could have done something similar for Outlook as well, but I did
not have this understanding of MAPI / Outlook at that time. Also the
human readnable XML base protocol has clarified many things and hence
this enlightnment.