Handle phone labels properly when construct EXContact object
As Exchange does not support custom labels for phone numbers we
store any customisations to the custom field. When we fetch a new
contact from an exchange store containing such labels stashed in
custom we need to pull that out and rest the contact fields
appropritely.

We also do this for fax labels. We are yet to do this for email
labels