Add ability to save the sync state from server to state.json
If we want to use the exchnage sync state support for synching we will need
abilityt to store collection level sync state inthe state.json

With this commit we add the requried support as well use that infra
to retreive and store the state during sync operations.

Note that there is a problem with this approach, as the exchange
sync state cannot be fetched after a push to the server... making
it virtually useless. Need to figure out the best way to work aroud
this...