Reimplement exchange prep_sync_lists ourselves
As explained in earlier commits, the exchange EWS operations only
allow us to fetch changes from the client. It is impossible to use
the syncstate infra to push changes from client to server.

So with this commit we fallback to the manaual generation of the
sync lists - by first fethcing all the items, and then looking that
up against the items list from the last sync.

It works. :)