Allow certain types of structured BBDB phone number with dot
BBDB seems liberal about what it allows to remain in the file
and parses as a structured phone number. For e.g. - (["Mobile" 0 0 0 60.8])
is a valid US number, athough the .8 part of the number is not
really displayed on the front end.

ASynK used to barf on such stuff. Now this is fixed; we are more
liberal in what we allow.