Fix BBDB sync to use get_itemid_from_synctags at folder level
Previously we were doing some hacky stuff in the contact
constructor that involved finding a sync label of the right
dbid. However with the move to Collections and with the ability
to sync between two dbs of same type this was not working any more
we also needed the profile name to do this, and it was getting
more hacky to this processing inside the constructor. With this
change we just do it at the higher level where we have all the
context to do this properly.

With this fix we take care of BBDB. The other DBs will follow in
time.