Allow --conflict-resolve to be set as 1 or 2
Earlier the argument only used to be the dbid - and as we did not allow
same-to-same synching, it was fine. With the present refactoring
we need to uniquely identify which of the two they are referring to.

With this commit we seamlessly migrate the identification of cr
flag from dbid to either 1 or 2. Existing state.json files will
work seamlessly as none of them would have issues with teh dbid
specification. Newly created profiles - whether same to same or
unique dbid syncs will all use 1 or 2 to specify the direction of
conflict resolution