Use oAuth Service Account to try and access contact list.
This is not working; but committing existing code so we can refer
back to what was not working.

Basically service accounts cannot refer to the user's contacts.

It would have been great if this had worked beacuase the service
accoutn client id and key can be used as the username and
password in all of the ASynK code without further messing.

But looksl ike this is not possible and we will have to move to
full blown oAuth. And that f-ing sucks.