BBDB notes_map web_home_re / web_work_re were not being used
This fixes #103, but it is only a hack. Need to rethink
the original idea of allowing regexp in the notes_map, or at least
make it a list of two values - a regexp to match at read time
and a pattern string to generate a label to write out.

At any rate the present hacky fix will work for those who are
using a single website per contact.