Take care of write time details related to BBDB file format ver 9 support
Ensure that the created and updated fields are written in the
correct format for ver 9 if that is the database version we found.

Also if ASynK is writing out to a new or blank BBDB database, then
the default format version will be 9 from here on. There is no
easy way to change that except fixing it in the code (but it is a
one line code change, easily indicated by the current commit diffs)